extracted helper functions to formatDateStamp.ts with documentation
relates #77
